Restructure to take out the base TreeModel class from the prim spec editor
