+++ usd_qtpy/tree/simpletree.py
@@ -1,0 +1,151 @@
+from PySide2 import QtCore
+
+
+class TreeModel(QtCore.QAbstractItemModel):
+    Columns = list()
+    ItemRole = QtCore.Qt.UserRole + 1
+
+    def __init__(self, parent=None):
+        super(TreeModel, self).__init__(parent)
+        self._root_item = Item()
+
+    def rowCount(self, parent=None):
+        if parent is None or not parent.isValid():
+            parent_item = self._root_item
+        else:
+            parent_item = parent.internalPointer()
+        return parent_item.childCount()
+
+    def columnCount(self, parent):
+        return len(self.Columns)
+
+    def data(self, index, role):
+        if not index.isValid():
+            return None
+
+        if role == QtCore.Qt.DisplayRole:
+            item = index.internalPointer()
+            column = index.column()
+
+            key = self.Columns[column]
+            return item.get(key, None)
+
+        if role == self.ItemRole:
+            return index.internalPointer()
+
+    def setColumns(self, keys):
+        assert isinstance(keys, (list, tuple))
+        self.Columns = keys
+
+    def headerData(self, section, orientation, role):
+
+        if role == QtCore.Qt.DisplayRole:
+            if section < len(self.Columns):
+                return self.Columns[section]
+        return
+
+        return super(TreeModel, self).headerData(section, orientation, role)
+
+    def flags(self, index):
+        flags = QtCore.Qt.ItemIsEnabled
+
+        item = index.internalPointer()
+        if item.get("enabled", True):
+            flags |= QtCore.Qt.ItemIsSelectable
+
+        return flags
+
+    def parent(self, index):
+
+        item = index.internalPointer()
+        parent_item = item.parent()
+
+        # If it has no parents we return invalid
+        if parent_item == self._root_item or not parent_item:
+            return QtCore.QModelIndex()
+
+        return self.createIndex(parent_item.row(), 0, parent_item)
+
+    def index(self, row, column, parent=None):
+        """Return index for row/column under parent"""
+
+        if parent is None or not parent.isValid():
+            parent_item = self._root_item
+        else:
+            parent_item = parent.internalPointer()
+
+        child_item = parent_item.child(row)
+        if child_item:
+            return self.createIndex(row, column, child_item)
+        else:
+            return QtCore.QModelIndex()
+
+    def add_child(self, item, parent=None):
+        if parent is None:
+            parent = self._root_item
+
+        parent.add_child(item)
+
+    def column_name(self, column):
+        """Return column key by index"""
+
+        if column < len(self.Columns):
+            return self.Columns[column]
+
+    def clear(self):
+        self.beginResetModel()
+        self._root_item = Item()
+        self.endResetModel()
+
+
+class Item(dict):
+    """An item that can be represented in a tree view using `TreeModel`.
+
+    The item can store data just like a regular dictionary.
+
+    >>> data = {"name": "John", "score": 10}
+    >>> item = Item(data)
+    >>> assert item["name"] == "John"
+
+    """
+
+    def __init__(self, data=None):
+        super(Item, self).__init__()
+
+        self._children = list()
+        self._parent = None
+
+        if data is not None:
+            assert isinstance(data, dict)
+            self.update(data)
+
+    def childCount(self):
+        return len(self._children)
+
+    def child(self, row):
+
+        if row >= len(self._children):
+            log.warning("Invalid row as child: %s", row)
+            return
+
+        return self._children[row]
+
+    def children(self):
+        return self._children
+
+    def parent(self):
+        return self._parent
+
+    def row(self):
+        """
+        Returns:
+             int: Index of this item under parent"""
+        if self._parent is not None:
+            siblings = self.parent().children()
+            return siblings.index(self)
+        return -1
+
+    def add_child(self, child):
+        """Add a child to this item"""
+        child._parent = self
+        self._children.append(child)
